feat(base): Moving out legacy common utils
