+++ module/rest/__index__.ts
@@ -26,4 +26,5 @@ export * from './src/types';
 export * from './src/util/param';
 export * from './src/util/route';
 export * from './src/util/mime';
+export * from './src/util/common';
 export * from './src/util/route-check';
