+++ module/rest/src/util/common.ts
@@ -1,0 +1,106 @@
+type List<T> = T[] | readonly T[];
+type OrderedState<T> = { after?: List<T>, before?: List<T>, key: T };
+
+export class RestCommonUtil {
+
+  static #match<T, K extends unknown[]>(
+    rules: { value: T, positive: boolean }[],
+    compare: (rule: T, ...compareInput: K) => boolean,
+    unmatchedValue: boolean,
+    ...input: K
+  ): boolean {
+    for (const rule of rules) {
+      if (compare(rule.value, ...input)) {
+        return rule.positive;
+      }
+    }
+    return unmatchedValue;
+  }
+
+  static #allowDenyRuleInput<T>(
+    rule: (string | T | [value: T, positive: boolean] | [value: T]),
+    convert: (inputRule: string) => T
+  ): { value: T, positive: boolean } {
+    return typeof rule === 'string' ?
+      { value: convert(rule.replace(/^!/, '')), positive: !rule.startsWith('!') } :
+      Array.isArray(rule) ?
+        { value: rule[0], positive: rule[1] ?? true } :
+        { value: rule, positive: true };
+  }
+
+  static #buildEdgeMap<T, U extends OrderedState<T>>(items: List<U>): Map<T, Set<T>> {
+    const edgeMap = new Map(items.map(x => [x.key, new Set(x.after ?? [])]));
+
+    // Build out edge map
+    for (const input of items) {
+      for (const bf of input.before ?? []) {
+        if (edgeMap.has(bf)) {
+          edgeMap.get(bf)!.add(input.key);
+        }
+      }
+      const afterSet = edgeMap.get(input.key)!;
+      for (const el of input.after ?? []) {
+        afterSet.add(el);
+      }
+    }
+    return edgeMap;
+  }
+
+  /**
+   * Produces a satisfied ordering for a list of orderable elements
+   */
+  static ordered<T, U extends OrderedState<T>>(items: List<U>): U[] {
+    const edgeMap = this.#buildEdgeMap<T, U>(items);
+
+    // Loop through all items again
+    const keys: T[] = [];
+    while (edgeMap.size > 0) {
+
+      // Find node with no dependencies
+      const key = [...edgeMap].find(([, after]) => after.size === 0)?.[0];
+      if (!key) {
+        throw new Error(`Unsatisfiable dependency: ${[...edgeMap.keys()]}`);
+      }
+
+      // Store, and remove
+      keys.push(key);
+      edgeMap.delete(key);
+
+      // Remove node from all other elements in `all`
+      for (const [, rem] of edgeMap) {
+        rem.delete(key);
+      }
+    }
+
+    const inputMap = new Map(items.map(x => [x.key, x]));
+    return keys.map(k => inputMap.get(k)!);
+  }
+
+  /**
+   * Simple check against allow/deny rules
+   * @param rules
+   */
+  static allowDenyMatcher<T, K extends unknown[]>(
+    rules: string | (string | T | [value: T, positive: boolean])[],
+    convert: (rule: string) => T,
+    compare: (rule: T, ...compareInput: K) => boolean,
+    cacheKey?: (...keyInput: K) => string
+  ): (...input: K) => boolean {
+
+    const rawRules = (Array.isArray(rules) ? rules : rules.split(/\s*,\s*/g));
+    const convertedRules = rawRules.map(rule => this.#allowDenyRuleInput(rule, convert));
+    const unmatchedValue = !convertedRules.some(r => r.positive);
+
+    if (convertedRules.length) {
+      if (cacheKey) {
+        const cache: Record<string, boolean> = {};
+        return (...input: K) =>
+          cache[cacheKey(...input)] ??= this.#match(convertedRules, compare, unmatchedValue, ...input);
+      } else {
+        return (...input: K) => this.#match(convertedRules, compare, unmatchedValue, ...input);
+      }
+    } else {
+      return () => true;
+    }
+  }
+}
