+++ module/rest/src/util/mime.ts
@@ -1,6 +1,5 @@
-import { Util } from '@travetto/base';
-
 import { ContentType } from '../types';
+import { RestCommonUtil } from './common';
 
 /**
  * Utils for checking mime patterns
@@ -26,7 +25,7 @@ export class MimeUtil {
    * Build matcher
    */
   static matcher(rules: string[] | string = []): (contentType: string) => boolean {
-    return Util.allowDenyMatcher<RegExp, [string]>(
+    return RestCommonUtil.allowDenyMatcher<RegExp, [string]>(
       rules,
       this.#convert.bind(this),
       (regex, mime) => regex.test(mime),
