+++ module/rest/src/util/route-check.ts
@@ -1,8 +1,7 @@
-import { Util } from '@travetto/base';
-
 import { RouteApplies } from '../interceptor/types';
 import { ControllerConfig } from '../registry/types';
 import { RouteConfig } from '../types';
+import { RestCommonUtil } from './common';
 
 interface RouteRule {
   sub: string | RegExp;
@@ -40,7 +39,7 @@ export class RouteCheckUtil {
    *   It is intended to be used during controller setup to determine an interceptors inclusion in the exposed endpoint.
    */
   static matcher(allowDeny: string[]): RouteApplies {
-    return Util.allowDenyMatcher<RouteRule, Parameters<RouteApplies>>(
+    return RestCommonUtil.allowDenyMatcher<RouteRule, Parameters<RouteApplies>>(
       allowDeny,
       this.#convertRule.bind(this),
       this.#compareRule.bind(this),
