+++ module/rest/test/common-util.ts
@@ -1,0 +1,47 @@
+import assert from 'node:assert';
+
+import { Test, Suite } from '@travetto/test';
+
+import { RestCommonUtil } from '../src/util/common';
+
+@Suite()
+export class RestCommonUtilTest {
+
+  @Test()
+  orderDependents() {
+    const items = [
+      {
+        key: 'first'
+      },
+      {
+        after: ['first', 'fourth'],
+        key: 'fifth'
+      },
+      {
+        after: ['first'],
+        key: 'third'
+      },
+      {
+        after: ['first'],
+        key: 'second'
+      },
+      {
+        after: ['first', 'second'],
+        key: 'fourth'
+      },
+      {
+        after: ['fifth'],
+        key: 'sixth'
+      }
+    ] as const;
+
+    const ordered = RestCommonUtil.ordered(items);
+    assert.deepStrictEqual(ordered.map(x => x.key), ['first', 'third', 'second', 'fourth', 'fifth', 'sixth']);
+
+    const ordered2 = RestCommonUtil.ordered([
+      { key: 'tenth', before: ['second'] },
+      ...items
+    ]);
+    assert.deepStrictEqual(ordered2.map(x => x.key), ['tenth', 'first', 'third', 'second', 'fourth', 'fifth', 'sixth']);
+  }
+}
